Fix: Resolve build errors from shared audio handling changes

+++ app/src/main/java/com/drgraff/speakkey/data/UploadTask.java
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey;
 public class UploadTask {
 
     @PrimaryKey(autoGenerate = true)
-    public int id;
+    public long id; // Changed from int to long
 
     public String filePath;
     public String uploadType; // e.g., "AUDIO_TRANSCRIPTION", "PHOTO_VISION"
